fix: Now fetch bills without interceptions
File interception did not work on iphone (the file was still displayed
on on the phone and blocked the rest of the execution of the konnector.

I found a way to download the files directly without clicking on the
links, which is faster and more reliable. And since downloadInWorker is
used, file will be downloaded only when needed.